+++ priv/templates/drab/drab.js
@@ -113,10 +113,6 @@
       };
       this.channel.push("event", message);
     },
-    run_handler: function(event_name, event_handler, payload, execute_after) {
-      Drab.error("Drab.run_handler() is depreciated. Please use Drab.exec_elixir() instead");
-      this.exec_elixir(event_handler, payload, execute_after);
-    },
     error: function (message) {
       console.log("[drab] " + message);
       if (this.channel) {