+++ test/support/web/commanders/share_commander.ex
@@ -13,7 +13,6 @@ defmodule DrabTestApp.ShareCommander do
   end
 
   defhandler defined_handler(socket, _sender) do
-    # IO.inspect(socket)
     set_prop(socket, ".spaceholder1", innerText: "changed globally")
     poke(socket, text: "set globally", bgcolor: "grey", color: "white")
   end